fix double-load problem
